requirements.txt
# Core dependencies for Automated Marker Detection Using YOLO
# Python 3.7+

# Deep Learning
torch>=1.7.0
torchvision>=0.8.0
ultralytics>=8.0.0  # For YOLOv8

# Computer Vision
opencv-python>=4.1.0
opencv-contrib-python>=4.1.0
Pillow>=7.1.0
scikit-image>=0.18.0

# Scientific Computing
numpy>=1.20.0
scipy>=1.4.0
pandas>=1.1.0

# Visualization
matplotlib>=3.2.0
seaborn>=0.11.0

# Utilities
tqdm>=4.64.0
pyyaml>=5.3.0

# Model Analysis
tensorboard>=2.4.0
thop>=0.1.1  # For FLOPs calculation

# Optional: For advanced features
# wandb>=0.12.0  # Weights & Biases logging
# clearml>=1.0.0  # ClearML logging
